+++ client/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added quotes around search strings containing only stop words ([#373](https://github.com/project-lux/lux-frontend/issues/373)).
+
 ### Changed
 
 - Archive Hierarchies display mixed Collections and Objects results (rather than Collections first, then Objects) ([#458](https://github.com/project-lux/lux-frontend/issues/458)).